+++ packages/soda-hooks/src/hooks/useBrowser.ts
@@ -1,0 +1,10 @@
+import { useMemo } from "react"
+
+import { getParser, Parser } from "bowser"
+
+export type BrowserParser = Parser.Parser
+
+export function useBrowser(): BrowserParser {
+    const browser = useMemo(() => getParser(window.navigator.userAgent), [])
+    return browser
+}