+++ packages/soda-heroui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # soda-tanstack-form
 
+## 0.12.19
+
+### Patch Changes
+
+- Updated dependencies
+    - soda-hooks@6.17.0
+
+## 0.12.18
+
+### Patch Changes
+
+- fix: 移除弃用的 baseUrl 字段
+- Updated dependencies
+    - soda-tanstack-form@0.3.7
+    - deepsea-tools@5.47.11
+    - soda-hooks@6.16.24
+    - soda-type@6.7.7
+
 ## 0.12.17
 
 ### Patch Changes