+++ packages/soda-next/CHANGELOG.md
@@ -1,5 +1,21 @@
 # soda-next
 
+## 0.7.30
+
+### Patch Changes
+
+- Updated dependencies
+    - soda-hooks@6.17.0
+
+## 0.7.29
+
+### Patch Changes
+
+- fix: 移除弃用的 baseUrl 字段
+- Updated dependencies
+    - deepsea-tools@5.47.11
+    - soda-hooks@6.16.24
+
 ## 0.7.28
 
 ### Patch Changes